Modifying Response json

+++ workforce-planning/src/main/java/com/frauas/workforce_planning/model/entity/ExternalEmployee.java
@@ -1,14 +1,21 @@
 package com.frauas.workforce_planning.model.entity;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
@@ -50,7 +57,7 @@ public class ExternalEmployee {
     private Double evaluationScore;
 
     @Column(name = "experience_years")
-    private Float experienceYears;
+    private Integer experienceYears;
 
     @Column(name = "wage_per_hour")
     private Double wagePerHour;
